InputGen: introduce variable space (#3)
Summary:

This diff introduces the VariableSpace class.

A VariableSpace needs to be initialized with a variable type. Its purpose is to describe a space of values for that variable type. This space can be discrete (like a Python set) or it can be a union of disjoint intervals of real numbers (only for suitable variable types).

Reviewed By: digantdesai

Differential Revision: D51816908