InputGen: introduce variable space (#3)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/FACTO/pull/3

This diff introduces the VariableSpace class.

A VariableSpace needs to be initialized with a variable type. Its purpose is to describe a space of values for that variable type. This space can be discrete (like a Python set) or it can be a union of disjoint intervals of real numbers (only for suitable variable types).

Reviewed By: digantdesai

Differential Revision: D51816908

fbshipit-source-id: e7bbfc26847bc3b1028dd15a05b441263a8497bc