InputGen: introduce solvable variable
Summary:
This diff introduces the SolvableVariable class, responsible for solving the space of values of a variable subjected to constraints.

The supported constraints are:
- Eq (equal to)
- Ne (not equal to)
- In (in list of values)
- NotIn (not in list of values)
- Le (less than or equal to)
- Lt (less than)
- Ge (greater than or equal to)
- Gt (greater than)

Note: A tentative future constraint not included here yet will be St (such that)

A SolvableVariable needs to be initialized with the variable type. It maintains a space of values for that variable type (VariableSpace object).
The result of applying multiple constraints to a SolvableVariable, is the conjunction of those constraints. Consequently, every constraint reduces or maintains its space. We currently don't support constraint disjunctions.

Sample usage:
```
v = SolvableVariable(float)
v.Gt(0.5)
v.Le(4.5)
v.NotIn([1.5, 2.5])
```

Reviewed By: SS-JIA

Differential Revision: D51868719

fbshipit-source-id: 51f740358561930b40e0b11c210228c0b24fc6f3